updates
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@6391 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 8c36e17f139379f6d0a103f8148a7b908e0e70a1

+++ Boole/Rich/RichReadout/src/RichPixelReadout.h
@@ -28,7 +28,7 @@ public:
   void setShape_flat ( RichShape_flat * ) ;
   void setNoisifier( RichNoisifier * ) ;
   void setADC( RichFrontEndDigitiser *) ;
-  void setBaseLine ( unsigned int index, double ) ;
+  void setBaseLine ( unsigned long index, double ) ;
   void setBaseLine ( double ) ;  //! same value for all index
   void setSigmaElecNoise ( double ) ;
   void setBinOfMaximum ( int ) ;
@@ -110,7 +110,7 @@ void RichPixelReadout::setADC (RichFrontEndDigitiser* p )
   m_frontEnd = p;
 }
 
-inline void RichPixelReadout::setBaseLine (unsigned int, double d)
+inline void RichPixelReadout::setBaseLine (unsigned long, double d)
 {
   m_baseline = d ;
 }
